+++ src/IOEither.ts
@@ -11,10 +11,10 @@ import { Bifunctor2 } from './Bifunctor'
 import { Compactable2C } from './Compactable'
 import * as E from './Either'
 import { Filterable2C } from './Filterable'
-import { bind_, flow, identity, Lazy, pipe, Predicate, Refinement, tuple } from './function'
+import { bind__, flow, identity, Lazy, pipe, Predicate, Refinement, tuple } from './function'
 import { bindTo_, Functor2 } from './Functor'
 import * as I from './IO'
-import { chainFirst_, Monad2 } from './Monad'
+import { bind_, chainFirst_, Monad2 } from './Monad'
 import { MonadIO2 } from './MonadIO'
 import { MonadThrow2 } from './MonadThrow'
 import { Monoid } from './Monoid'
@@ -674,16 +674,12 @@ export const bindTo: <N extends string>(name: N) => <E, A>(fa: IOEither<E, A>) =
 /**
  * @since 3.0.0
  */
-export const bindW = <N extends string, A, D, B>(
+export const bindW: <N extends string, A, D, B>(
   name: Exclude<N, keyof A>,
   f: (a: A) => IOEither<D, B>
-): (<E>(fa: IOEither<E, A>) => IOEither<D | E, { [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
-  chainW((a) =>
-    pipe(
-      f(a),
-      map((b) => bind_(a, name, b))
-    )
-  )
+) => <E>(fa: IOEither<E, A>) => IOEither<D | E, { [K in keyof A | N]: K extends keyof A ? A[K] : B }> =
+  /*#__PURE__*/
+  bind_(Monad) as any
 
 /**
  * @since 3.0.0
@@ -705,7 +701,7 @@ export const apSW = <A, N extends string, D, B>(
   fb: IOEither<D, B>
 ): (<E>(fa: IOEither<E, A>) => IOEither<D | E, { [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
   flow(
-    map((a) => (b: B) => bind_(a, name, b)),
+    map((a) => (b: B) => bind__(a, name, b)),
     apW(fb)
   )