+++ src/ReaderTask.ts
@@ -3,10 +3,10 @@
  */
 import { Applicative2 } from './Applicative'
 import { apFirst_, apSecond_ } from './Apply'
-import { bind_, flow, identity, pipe, tuple } from './function'
+import { bind__, flow, identity, pipe, tuple } from './function'
 import { bindTo_, Functor2 } from './Functor'
 import { IO } from './IO'
-import { chainFirst_, Monad2 } from './Monad'
+import { bind_, chainFirst_, Monad2 } from './Monad'
 import { MonadTask2 } from './MonadTask'
 import { Monoid } from './Monoid'
 import * as R from './Reader'
@@ -319,16 +319,12 @@ export const bindTo: <N extends string>(name: N) => <R, A>(fa: ReaderTask<R, A>)
 /**
  * @since 3.0.0
  */
-export const bindW = <N extends string, A, Q, B>(
+export const bindW: <N extends string, A, Q, B>(
   name: Exclude<N, keyof A>,
   f: (a: A) => ReaderTask<Q, B>
-): (<R>(fa: ReaderTask<R, A>) => ReaderTask<Q & R, { [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
-  chainW((a) =>
-    pipe(
-      f(a),
-      map((b) => bind_(a, name, b))
-    )
-  )
+) => <R>(fa: ReaderTask<R, A>) => ReaderTask<Q & R, { [K in keyof A | N]: K extends keyof A ? A[K] : B }> =
+  /*#__PURE__*/
+  bind_(Monad) as any
 
 /**
  * @since 3.0.0
@@ -350,7 +346,7 @@ export const apSW = <A, N extends string, Q, B>(
   fb: ReaderTask<Q, B>
 ): (<R>(fa: ReaderTask<R, A>) => ReaderTask<Q & R, { [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
   flow(
-    map((a) => (b: B) => bind_(a, name, b)),
+    map((a) => (b: B) => bind__(a, name, b)),
     apW(fb)
   )