+++ src/ReadonlyNonEmptyArray.ts
@@ -10,10 +10,10 @@ import { Eq } from './Eq'
 import { Extend1 } from './Extend'
 import { Foldable1 } from './Foldable'
 import { FoldableWithIndex1 } from './FoldableWithIndex'
-import { bind_, Endomorphism, flow, Lazy, pipe, Predicate, Refinement, tuple } from './function'
+import { bind__, Endomorphism, flow, Lazy, pipe, Predicate, Refinement, tuple } from './function'
 import { bindTo_, Functor1 } from './Functor'
 import { FunctorWithIndex1 } from './FunctorWithIndex'
-import { Monad1 } from './Monad'
+import { bind_, Monad1 } from './Monad'
 import { none, Option, some } from './Option'
 import { Ord } from './Ord'
 import * as RA from './ReadonlyArray'
@@ -735,16 +735,12 @@ export const bindTo: <N extends string>(
 /**
  * @since 3.0.0
  */
-export const bind = <N extends string, A, B>(
+export const bind: <N extends string, A, B>(
   name: Exclude<N, keyof A>,
   f: (a: A) => ReadonlyNonEmptyArray<B>
-): ((fa: ReadonlyNonEmptyArray<A>) => ReadonlyNonEmptyArray<{ [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
-  chain((a) =>
-    pipe(
-      f(a),
-      map((b) => bind_(a, name, b))
-    )
-  )
+) => (fa: ReadonlyNonEmptyArray<A>) => ReadonlyNonEmptyArray<{ [K in N | keyof A]: K extends keyof A ? A[K] : B }> =
+  /*#__PURE__*/
+  bind_(Monad)
 
 // -------------------------------------------------------------------------------------
 // pipeable sequence S
@@ -758,7 +754,7 @@ export const apS = <A, N extends string, B>(
   fb: ReadonlyNonEmptyArray<B>
 ): ((fa: ReadonlyNonEmptyArray<A>) => ReadonlyNonEmptyArray<{ [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
   flow(
-    map((a) => (b: B) => bind_(a, name, b)),
+    map((a) => (b: B) => bind__(a, name, b)),
     ap(fb)
   )